rtpst2022-1-fecenc: fix input seqnum check
We need to cast the incremented last seqnum to guint16 for
consistent checks on wraparound

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/-/merge_requests/770>

+++ gst/rtpmanager/gstrtpst2022-1-fecenc.c
@@ -389,7 +389,8 @@ gst_rtpst_2022_1_fecenc_sink_chain (GstPad * pad, GstObject * parent,
   }
 
   if (enc->last_media_seqnum_set
-      && enc->last_media_seqnum + 1 != gst_rtp_buffer_get_seq (&rtp)) {
+      && (guint16) (enc->last_media_seqnum + 1) !=
+      gst_rtp_buffer_get_seq (&rtp)) {
     GST_ERROR_OBJECT (enc, "consecutive sequence numbers are required");
     goto error;
   }
